Added api layer and di part added

+++ domain/search/src/main/java/com/app/test/domain/search/models/Hotel.kt
@@ -11,6 +11,6 @@ data class Hotel(
 )
 
 data class HotelGeoLocation(
-    val lat: Long,
-    val long: Long,
+    val lat: Double,
+    val long: Double,
 )
